pyproject.toml
# Configuration existante pour les outils de build
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Configuration Poetry (à ajouter)
[tool.poetry]
name = "git-commit-simplifier"
version = "0.1.3"
description = "A tool to simplify Git commit message creation by analyzing changes"
authors = ["Christopher Dato <christopherdato08@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Chris000888/git-commit-simplifier"
license = "MIT"
packages = [{include = "git_commit_simplifier", from = "src"}]
keywords = ["git", "commit", "message", "automation", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

# Dépendances du projet
[tool.poetry.dependencies]
python = ">=3.8,<4.0"
gitpython = "^3.1.40"
click = "^8.1.7"
colorama = "^0.4.6"
prompt_toolkit = "^3.0.36"

# Dépendances de développement
[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"

# Points d'entrée CLI
[tool.poetry.scripts]
git-commit-simplifier = "git_commit_simplifier.cli:main"
git_commit_simplifier = "git_commit_simplifier.cli:main"
gitcommitsimplifier = "git_commit_simplifier.cli:main"
gcs = "git_commit_simplifier.cli:main"

# Conserver les configurations existantes pour Black
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

# Conserver les configurations existantes pour isort
[tool.isort]
profile = "black"
multi_line_output = 3
